540 fetch and provide manifests from p2p
